feat: Replace Material Icons with Material Symbols
This commit replaces all instances of the default Material Icons (the `Icons` class) with icons from the `material_symbols_icons` package (the `Symbols` class).

This change was made to use the latest and more comprehensive set of icons provided by the Material Symbols package.

The `_outlined` and `_rounded` variants of the icons were replaced with their base counterparts from the new package.
The `const` keyword was removed from `Icon` widgets where the icon from the `Symbols` class was used, as these are not constant values.